All fetch called made from TempLogin.tsx
