+++ frontend/src/components/tempLogin/TempPlayerList.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react'
+import { useEffect } from 'react'
 import TempPlayerLi from './TempPlayerLi'
 import styles from './TempPlayerList.module.css'
 
@@ -7,34 +7,22 @@ interface Player {
     email: string
     avatarUrl: string
 }
-async function getUsers() {
-    try {
-        const response = await fetch('http://localhost:8080/api/player')
 
-        if (!response.ok) {
-            throw new Error('Failed to fetch users')
-        }
-
-        const users = await response.json()
-        return users
-    } catch (error) {
-        console.error(error)
-    }
+interface TempPlayerListProps {
+    getPlayers: () => void
+    players: Player[]
 }
 
-const TempPlayerList = () => {
-    const [players, setPlayers] = useState([])
-
+const TempPlayerList = (props: TempPlayerListProps) => {
     useEffect(() => {
-        getUsers().then((users) => {
-            setPlayers(users)
-            console.log(users)
-        })
+        props.getPlayers()
     }, [])
 
     let content: JSX.Element[] | JSX.Element = <p>No Player Found!</p>
-    if (players.length > 0) {
-        content = players.map((player: Player) => <TempPlayerLi key={player.login} player={player}></TempPlayerLi>);
+    if (props.players.length > 0) {
+        content = props.players.map((player: Player) => (
+            <TempPlayerLi key={player.login} player={player}></TempPlayerLi>
+        ))
     }
 
     return (
